backend/pyproject.toml
[project]
name = "mabara-voucher"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiomysql>=0.2.0",
    "bcrypt>=4.3.0",
    "bonsai>=1.5.3",
    "fastapi[standard]>=0.115.12",
    "icecream>=2.1.4",
    "pydantic-settings>=2.9.1",
    "python-dotenv>=1.1.0",
    "python-jose>=3.4.0",
    "pytz>=2025.2",
    "sqlmodel>=0.0.24",
]
